requirements.txt
# For frida-ios-dump
frida-tools

# For this script
python-dotenv
discord.py
requests
paramiko
